+++ lib/darbiw.dart
@@ -1,4 +1,5 @@
-library binary;
+library;
 
 export 'src/binary_writer.dart';
 export 'src/binary_reader.dart';
+export 'src/annotations.dart';