+++ lib/src/generator.dart
@@ -4,21 +4,28 @@ import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/dart/element/nullability_suffix.dart';
 import 'package:analyzer/dart/element/type.dart';
 import 'package:build/build.dart';
-import 'package:darbiw/annotations.dart';
+import 'package:darbiw/darbiw.dart';
 import 'package:source_gen/source_gen.dart';
+
+final _datetimeChecker = TypeChecker.fromRuntime(DateTime);
+final _durationChecker = TypeChecker.fromRuntime(Duration);
 
 final class BinaryGenerator extends GeneratorForAnnotation<Binary> {
   @override
-  FutureOr<String> generateForAnnotatedElement(Element element, ConstantReader annotation, BuildStep buildStep) {
+  FutureOr<String> generateForAnnotatedElement(
+      Element element, ConstantReader annotation, BuildStep buildStep) {
     final encoderBuffer = StringBuffer();
     final decoderBuffer = StringBuffer();
     if (element is! ClassElement) {
       throw "Binary annotation must be only added to classes.";
     }
 
-    final constructor = element.constructors.where((element) => element.name.isEmpty).firstOrNull;
+    final constructor = element.constructors
+        .where((element) => element.name.isEmpty)
+        .firstOrNull;
     if (constructor == null) {
-      print("ignoring ${element.name} because not default constructor is found.");
+      print(
+          "ignoring ${element.name} because not default constructor is found.");
       return "";
     }
 
@@ -31,7 +38,8 @@ final class BinaryGenerator extends GeneratorForAnnotation<Binary> {
     final encoders = <String>[];
     final decoders = <String>[];
     for (final parameter in constructor.parameters) {
-      final (writer, size) = _getNullabilityWriter(parameter.name, parameter.type);
+      final (writer, size) =
+          _getNullabilityWriter(parameter.name, parameter.type);
       bufferSize += size;
       encoders.add(writer);
 
@@ -45,21 +53,25 @@ final class BinaryGenerator extends GeneratorForAnnotation<Binary> {
         final writer = BinaryWriter($bufferSize);
         """);
     encoderBuffer.writeAll(encoders);
-    encoderBuffer.write("return writer.takeBytes(); }}"); // end of toBuffer. end of extension declx
+    encoderBuffer.write(
+        "return writer.takeBytes(); }}"); // end of toBuffer. end of extension declx
 
     // write XFromBuffer method
     decoderBuffer.write('@pragma("vm:prefer-inline")');
-    decoderBuffer.write("""${element.name} _${element.name}FromBuffer(Uint8List buffer) {
+    decoderBuffer
+        .write("""${element.name} _${element.name}FromBuffer(Uint8List buffer) {
       final reader = BinaryReader(buffer);
       return ${element.name}(""");
     decoderBuffer.writeAll(decoders);
-    decoderBuffer.write("); }"); // end of constructor - end of XFromBuffer method
+    decoderBuffer
+        .write("); }"); // end of constructor - end of XFromBuffer method
 
     return encoderBuffer.toString() + decoderBuffer.toString();
   }
 }
 
-(String writer, int bufferSize) _getNullabilityWriter(String paramName, DartType type) {
+(String writer, int bufferSize) _getNullabilityWriter(
+    String paramName, DartType type) {
   final bool isNullable = type.nullabilitySuffix == NullabilitySuffix.question;
   var (writer, bufferSize) = _getWriter(paramName, type, isNullable);
 
@@ -70,7 +82,8 @@ final class BinaryGenerator extends GeneratorForAnnotation<Binary> {
   return (writer, bufferSize);
 }
 
-(String writer, int bufferSize) _getWriter(String paramName, DartType type, [bool isNullable = false]) {
+(String writer, int bufferSize) _getWriter(String paramName, DartType type,
+    [bool isNullable = false]) {
   if (type.isDartCoreBool) {
     return ("writer.writeByte($paramName ? 1 : 0);", 1);
   } else if (type.isDartCoreDouble) {
@@ -78,16 +91,23 @@ final class BinaryGenerator extends GeneratorForAnnotation<Binary> {
   } else if (type.isDartCoreInt) {
     return ("writer.writeInt($paramName);", 8);
   } else if (type.isDartCoreNum) {
-    return ("$paramName is int ? writer.writeInt($paramName) : writer.writeDouble($paramName);", 8);
+    return (
+      "$paramName is int ? writer.writeInt($paramName) : writer.writeDouble($paramName);",
+      8
+    );
   } else if (type.isDartCoreString) {
     return ("writer.writeString($paramName);", 300);
   } else if (type.isDartCoreList || type.isDartCoreSet) {
     // get base type
     final baseType = (type as ParameterizedType).typeArguments.first;
-    final baseTypeIsNullable = baseType.nullabilitySuffix == NullabilitySuffix.question;
-    var (baseTypeWriter, size) = _getWriter("$paramName${isNullable ? '!' : ''}[i]${baseTypeIsNullable ? '!' : ''}", baseType);
+    final baseTypeIsNullable =
+        baseType.nullabilitySuffix == NullabilitySuffix.question;
+    var (baseTypeWriter, size) = _getWriter(
+        "$paramName${isNullable ? '!' : ''}[i]${baseTypeIsNullable ? '!' : ''}",
+        baseType);
     if (baseTypeIsNullable) {
-      baseTypeWriter = "$paramName${isNullable ? '!' : ''}[i] == null ? writer.writeNull() : $baseTypeWriter";
+      baseTypeWriter =
+          "$paramName${isNullable ? '!' : ''}[i] == null ? writer.writeNull() : $baseTypeWriter";
     }
 
     return (
@@ -130,7 +150,15 @@ for(final MapEntry(:key, :value) in $paramName${isNullable ? '!' : ''}.entries) 
       // bufferSize += 1; // maybe increase this for bigger enums
       return ("writer.writeByte($paramName${isNullable ? '!' : ''}.index);", 1);
     } else if (type.element is ClassElement) {
-      return ("writer.writeUint8List($paramName${isNullable ? '!' : ''}.toBuffer());", 512);
+      if (_datetimeChecker.isAssignableFrom(type.element)) {
+        return ("writer.writeInt($paramName.millisecondsSinceEpoch);", 8);
+      } else if (_durationChecker.isAssignableFrom(type.element)) {
+        return ("writer.writeInt($paramName.inMilliseconds);", 8);
+      }
+      return (
+        "writer.writeUint8List($paramName${isNullable ? '!' : ''}.toBuffer());",
+        512
+      );
     }
   }
 
@@ -176,6 +204,11 @@ String _getReader(DartType type) {
     if (type.element is EnumElement) {
       return "${type.getDisplayString(withNullability: false)}.values[reader.readByte()]";
     } else if (type.element is ClassElement) {
+      if (_datetimeChecker.isAssignableFrom(type.element)) {
+        return "${type.getDisplayString(withNullability: false)}.fromMillisecondsSinceEpoch(reader.readInt())";
+      } else if (_durationChecker.isAssignableFrom(type.element)) {
+        return "${type.getDisplayString(withNullability: false)}(milliseconds: reader.readInt())";
+      }
       return "${type.getDisplayString(withNullability: false)}.fromBuffer(reader.readUint8List())";
     }
   }