+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.1
+
+- Single export for all classes
+- Add support for `DateTime` and `Duration`
+
 ## 1.0.0
 
 - Initial version.